changed INSPEED type to long

+++ Ultrasonic/ultrasonic.h
@@ -8,7 +8,7 @@ class Ultrasonic
         void StartTrigger();
     public:
         const unsigned int CMSPEED = 34300;
-        const unsigned int INSPEED = 13511.8110236;
+        const unsigned long INSPEED = 13511.8110236;
         Ultrasonic(unsigned short t_pin, unsigned short e_pin);
         long Measure();
         long GetRawTime();
